Initial update to script to use new country code list and check before reloading

+++ dataactvalidator/scripts/load_country_codes.py
@@ -2,30 +2,32 @@ import os
 import logging
 
 import pandas as pd
-import numpy as np
 import boto3
 import datetime
 import json
 
+from dataactbroker.helpers.pandas_helper import check_dataframe_diff
+
 from dataactcore.config import CONFIG_BROKER
 from dataactcore.interfaces.db import GlobalDB
 from dataactcore.models.domainModels import CountryCode
+
 from dataactvalidator.health_check import create_app
 from dataactvalidator.scripts.loader_utils import clean_data, insert_dataframe
 
 logger = logging.getLogger(__name__)
 
 
-# Territories or freely associated states
-TERRITORIES_FREE_STATES = ["ASM", "XBK", "GUM", "XHO", "XJV", "XJA", "XKR", "XMW", "XNV", "MNP", "PRI", "XPL", "VIR",
-                           "XWK", "PLW", "FSM", "MHL"]
+def convert_bool_to_str(b_to_s):
+    return str(b_to_s)
 
 
-def load_country_codes(base_path):
+def load_country_codes(base_path, force_reload=False):
     """ Load Country Codes into the database.
 
         Args:
             base_path: directory that contains the domain values files.
+            force_reload: boolean to determine if reload should happen whether there are differences or not
     """
     now = datetime.datetime.now()
     metrics_json = {
@@ -48,29 +50,37 @@ def load_country_codes(base_path):
     with create_app().app_context():
         sess = GlobalDB.db().session
 
-        # for object class, delete and replace values
-        metrics_json['records_deleted'] = sess.query(CountryCode).delete()
-
         data = pd.read_csv(filename, dtype=str)
         metrics_json['records_provided'] = len(data.index)
         data = clean_data(
             data,
             CountryCode,
-            {"country_code": "country_code", "country_name": "country_name"},
+            {"country_code": 'country_code',
+             'country_name': 'country_name',
+             'territory_or_freely_associated_state': 'territory_free_state'},
             {}
         )
         # de-dupe
         data.drop_duplicates(subset=['country_code'], inplace=True)
         metrics_json['duplicates_dropped'] = metrics_json['records_provided'] - len(data.index)
-        # flag territories or freely associated states
-        data["territory_free_state"] = np.where(data["country_code"].isin(TERRITORIES_FREE_STATES), True, False)
-        # insert to db
-        table_name = CountryCode.__table__.name
-        num = insert_dataframe(data, table_name, sess.connection())
-        metrics_json['records_inserted'] = num
-        sess.commit()
 
-    logger.info('{} records inserted to {}'.format(num, table_name))
+        # compare to existing content in table
+        diff_found = check_dataframe_diff(data, CountryCode, 'country_code_id', ['country_code'],
+                                          lambda_funcs={'territory_free_state': convert_bool_to_str})
+
+        # insert to db if reload required
+        if force_reload or diff_found:
+            logger.info('Differences found or reload forced, reloading country_code table.')
+            # if there's a difference, clear out the old data before adding the new stuff
+            metrics_json['records_deleted'] = sess.query(CountryCode).delete()
+
+            num = insert_dataframe(data, CountryCode.__table__.name, sess.connection())
+            metrics_json['records_inserted'] = num
+            sess.commit()
+
+            logger.info('{} records inserted to country_code table'.format(num))
+        else:
+            logger.info('No differences found, skipping country_code table reload.')
 
     metrics_json['duration'] = str(datetime.datetime.now() - now)
